local_clock: Add output to flaky test (#1762)
This test is naturally a tiny bit flaky, as it depends on the passage of
real time. However in a first step to determine if we can improve it at
all, let's see what exactly is going wrong when it does go wrong. Add a
bit of output to the asserts so we can see why they fail. Part of
https://github.com/microsoft/openvmm/issues/1475.