hvlite_core: handle guest-initialiated reset in LoadedVm (#1767)
This change allows a `LoadedVm` to intercept a halt event and optionally
reset the VM when the guest requests it. This allows the guest to reboot
when using the ttrpc/grpc interface for OpenVMM. The configuration also
accepts a number of allowed restarts, which should make it easier to
allow extra restarts for certain guests in petri tests.