vmbus_server: fix a race condition when revoking and reoffering channels (#1815)
If a channel is explicitly revoked using `ChannelServerRequest::Revoke`
(used by both the OpenHCL relay and vmbusproxy), there is a small window
where the offer ID can be reused before the `ChannelServerRequest`
stream is dropped, at which point the revoke is attempted again. This
could end up revoking the newly offered channel, rather than the
original one.

This change updates the revoke logic to check the channel's sequence
number, so it will not revoke newer channels.