tdx: use apic id in reset vector (#1824)
Resolves #1818 . The kernel uses `APIC_IDs` to identify which AP is
being started, but the reset vector identifies the AP using the
`VCPU_INDEX`. This breaks startup on for any TD large enough to have an
`APIC_ID` which is not equal to the `VCPU_INDEX`.

This PR updates the reset vector to read the APIC ID msr with a
TDG.VP.VMCALL.