vmbus_server: revert to creating interrupts before channel open (#1850)
This change reverts the changes made by #649. While the model used by
that change was better, it was discovered that some channels, notably
some IC channels, begin communicating as soon as the device is notified
of the open. This leads to the possibility of the guest attempting to
send an interrupt before the synic event port was created, so the
interrupt would be lost.

This change therefore moves the creation of the guest-to-host event port
back to when the open request is received.

Note that this is not a straight-forward revert of the original commit,
as the code had diverged too much, particularly because the synic
handling for vmbus_relay was moved into vmbus_client. This is therefore
essentially a new change that reverts to the same functionality as
before.