vmbus_server: fix incorrect assumptions about proxy driver state (#1861)
The `proxyintegration` module made a number of assertions on the state
of the vmbusproxy driver. However, because handling server requests and
proxy driver requests is done in separate threads, various operations
could race with a channel being revoked. If that happens, ioctl calls
into the driver might fail with `ERROR_NOT_FOUND` because the channel
they were operating on no longer exists. Similarly, the channel may no
longer exist in the `channels` or `gpadls` maps.

This change fixes those assumptions and changes them to checks,
preventing a panic in case such a race occurs.