TDX: Fix a VTL 1-only typo, but move logic to only affect VTL 0 (#1859)
Masanamuthu spotted this copy/paste typo in vm/x86/x86defs/src/tdx.rs,
fix that up. However I think in the only place we set any TDX controls
today we actually only want those controls to apply to VTL 0, so move
that logic out of the for loop. This makes this PR a functional no-op,
but it fixes the typo for when the day comes that we do want to set a
control on VTL 1.