openhcl_boot, arm64: let `VBAR_EL1` entries use crash registers (#1874)
When the bootshim fails on arm64 in the low-level paths/ways, it'll just
spin. Enlighten the VBAR_EL1 entries to write to the crash registers to
have diagnostics if something goes wrong early in the early code, e.g.
when servicing.

Few other fixes included, seemed appropriate:
* do not use `weak _DYNAMIC` as that hides the issues with the toolchain
and leads to hard to diagnose failures,
* do not spin on relocation failures,
* install `VBAR_EL1` as early as possible,
* always inline the `fault` and `dead_loop` intrinsics.

Tested by inserting `brk` in various places in the asm. Observed the
crash report on the host.