vmbus_server: improve gpadl message error handling (#1860)
This change updates error handling for the `GpadlHeader` and `GpadlBody`
messages to ensure an error response is sent to the guest in the same
situations that Hyper-V's VMBus server would send one.

This change also moves the `channels::tests` module to a separate file
because it was getting unwieldy. The only new code in tests.rs is the
`test_gpadl_create_failure` function.