build_igvm: Separate release configuration from build profile (#1867)
This PR adds a new flag to local invocations of `xflowey build-igvm`,
`--release-cfg`. This flag controls various bits of configuration around
how the final IGVM file is produced, such as the json file used and the
features compiled in. This separates these configuration pieces from
`--release`, which now only control the build profile. This allows for
building an optimized binary with debug options still available. It also
allows for building an unoptimized output without debug options, but
this is weird, so it outputs a warning if you ask for it.

This PR then removes the debug configurations from the cvm-release json.
This means that running `xflowey build-igvm x64-cvm --release` locally
will still produce an optimized binary with the debug bits set, but it
means the binaries we ship to the OS will now have the debug bit unset
by default, as the publishing step is effectively running with
`--release --release-cfg`.