Repo: Update to the 2024 edition (#1425)
I have spot checked about 50 of the 300-ish hits the remaining lints
have, as well as manually searching for known potentially problematic
patterns, and am relatively certain that the drop order changes will not
affect us. Thus the last blocker is removed, and we can update. This
naturally introduced a few new minor lint issues, so those are fixed
here.

Closes #288