virt_*: The big error cleanup (#1799)
This PR does a number of things to significantly clean up error handling
through the multiple virt backends:

1. Remove the generic from virt::VpHaltReason, in favor of boxing

This allows us to chunk up our overly large error enums into much
smaller structs, as they no longer need to all be under the same type.

2. Remove the vast majority of uses of VpHaltReason::Hypervisor

The reasoning being that if the hypervisor is truly at fault we can not
recover and may be in a bad state. Most of these errors have been
changed to unwraps, with a few being changed to InvalidVmState, and even
fewer left alone.

3. Remove Result returns on functions that never fail

This significantly cleans up a number of trait definitions, and should
result in more compact code.

I think more can be done along these lines, but this took me a while as
it is so I decided to call it here for now. It's definitely possible
we'll want to walk this back in some of these places, very open to
feedback.