petri: unblock commands and use com3 for hyper-v kmsg logs when possible (#1176)
Using COM3 instead of diag_client allows us to get OpenHCL logs if it
crashes early in boot. Currently this only works on our TDX and SNP
runners, since our nested runners have a build of windows that is too
old and additional com ports are not yet supported on ARM. To make this
work, the powershell and hvc commands needed to be run in a separate
process (using the `unblock` function) to avoid blocking the process
doing the logging. Also fixes guest serial logs that fail to reconnect after
os error 231 (ubuntu 22.04 tests).