nvme_test: Added panic fault for admin commands (#1925)
Previously the openhcl test was using the `QueueFaultBehaviour::Drop`
fault in order to block `CREATE_IO_COMPLETION_QUEUE` commands. In the
test failure case this relied on the test timing out and failing. In
order to fail fast we can instead panic if such a command is seen. This
should reduce test run time in the failure case.