nvme: improve shadow doorbell support (#1582)
Currently, shadow doorbells are implemented in a way where there are
essentially two doorbell values: the shadow one and the "real" one.
Observed changes to the shadow value are reflected in the real value but
not vice versa. This causes problems when guests do not use, or only
inconsistently use, the shadow doorbells for a given queue.

The spec is somewhat ambiguous on this front, but it's clear from guest
behavior that the expectation is that the shadow doorbells and the real
doorbells remain in perfect sync, so that the actual doorbell value can
be read from the shadow doorbell. To support this, use the shadow
doorbell memory as the single source of truth for the doorbell register
value.

In addition to these changes, simplify the EventIdx handling to match
the strategy we use for virtio and vmbus devices--update EventIdx (thus
letting the guest know it needs to write the real doorbell on the next
request) only when polling for work and the queue is currently empty.
This is easier to reason about than the existing heuristic approach.

This does eliminate an optimization where we currently avoid updating
EventIdx when there are IOs in flight (thus increasing IO latency but
reducing the intercept rate), but a future change can easily add this
behavior back by avoiding polling (and instead just probing) a queue
when more than the desired number of IOs are in flight.

These changes are an improvement on their own, but they are also prep
work for supporting multiple submission queues per completion queue,
which is required by the NVMe spec and leveraged by Windows guests.