flowey: use release branches to download artifacts for servicing tests (#1918)
This PR is a slight re-work of #1897, which got reverted by #1915 to
quickly resolve a build break introduced by the original PR in the
internal mirror of openvmm.

The main difference between #1897 and this PR is that artifact
resolution is done at a later stage in the job which avoids having to
propagate pipeline variables unnecessarily throughout the job.

---------

Co-authored-by: Daman Mulye <daman.mulye@hotmail.com>