vmgstool: recover if no space for new encryption key (#1958)
Unify the logic for updating the VMGS encryption key by moving the logic
into a new function in `vmgs_impl`. OpenHCL already had the logic for
recovering when adding a new key and there was already two datastore
keys, but this error was encountered when using VmgsTool. This can occur
if the previous add key operation failed mid-way through and we did not
remove the old key.