vmm_tests: nvme keepalive with faulty identify:namespace after servicing  (#1977)
Adding a test for testing nvme_keepalive that provides a faulty response
to the IDENTIFY:NAMESPACE admin command and reports a changed size for
the already existing namespace. This test should be successful because
(for now) the nvme driver polls the namespaces after reboot. As long as
the reported size is not 0 the driver is able to make progress. Once AER
is implemented for the nvme driver this behavior is set to change and
the test will be updated accordingly.
I found that the previous nvme test had accidentally been duplicated. It
has been removed as part of this PR as well.