vmbus_server: support OpenHCL MNF on CVM (#1980)
When OpenHCL is running in a CVM, enabling MNF to be handled by OpenHCL
was not possible because the guest always sends host-visible pages for
MNF, and page protections for host-visible pages cannot be changed.

This change addresses this issue by allowing the VMBus server to supply
the monitor pages, controlled by a new feature flag. When this flag is
set, the server may ignore the guest-supplied monitor pages and allocate
its own, returning the new values in the `VersionResponse` message. This
allows VMBus to allocate private pages for which we can receive an
intercept to emulate writes and implement MNF behavior.

To allow switching behavior depending on whether MNF is handled by
OpenHCL or the relay host, this change updates how MNF is "disabled" if
MNF is handled by OpenHCL but the guest doesn't support server-supplied
monitor pages. In that case, the server no longer sends the monitor ID
for any channels, which prevents the guest from attempting to send
interrupts using MNF.

For now, server-supplied monitor pages are only supported by OpenHCL in
a CVM. In a regular OpenHCL VM, or an OpenVMM VM, this functionality is
not available.

Guests will have to be updated to support this functionality.

This change also restores the changes made in #1541 and reverted in
#1731, with some additional modifications.

Addresses #1711.