repo: Fix lints for Rust 1.90 (#2020)
The vast majority of these are a new lint that replaces manual modular
arithmetic with the is_multiple_of helper, which were all just applied
automatically. There were also a few cases of extra unneeded
parentheses, and one new allow(dead_code) in a protocol definitions file
since not all of it is used today.